Add Increment: Add Item Price
This version of easyLog has an additional attribute called "price" for all items. The user is able to / has to add a price for the item to be added. The price should be a number which is between 0 and 1,000,000,000 SGD inclusive. The item list will also display the prices for different items in the system.