Add Increment: Update Item Price
This version of easyLog can update the additional attribute called "price" for all items. The user is able to revise an item price any time at ease. Again, the price should be a number which is between 0 and 1,000,000,000 SGD inclusive. The item list will also display the revised prices for different items in the system.
(Zikun also adds the code base so that other fields can be updated with the same simple logic (e.g., stock).)